requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
fastapi-clerk-auth
pyjwt
